Fix wrong data type

+++ src/mesh.cpp
@@ -28,7 +28,7 @@ void Mesh::Draw(Shader shader) {
             number = to_string(heightNr++);
         }
 
-        shader.setFloat(("material." + name + number).c_str(), i);
+        shader.setInt(("material." + name + number).c_str(), i);
         // glUniform1i(glGetUniformLocation(shader.shaderProgram, (name + number).c_str()), i);
         glBindTexture(GL_TEXTURE_2D, texture[i].id);
     }
